Refactor storage model to work with native Realm
