+++ ios/App/Shared/models/PlayerState.swift
@@ -1,0 +1,15 @@
+//
+//  PlayerState.swift
+//  App
+//
+//  Created by Ron Heft on 8/16/22.
+//
+
+import Foundation
+
+enum PlayerState: Codable {
+    case IDLE
+    case BUFFERING
+    case READY
+    case ENDED
+}
