+++ ios/App/Shared/models/server/MediaProgress.swift
@@ -1,0 +1,58 @@
+//
+//  MediaProgress.swift
+//  App
+//
+//  Created by Ron Heft on 8/16/22.
+//
+
+import Foundation
+import RealmSwift
+
+class MediaProgress: EmbeddedObject, Codable {
+    @Persisted var id: String = ""
+    @Persisted var libraryItemId: String = ""
+    @Persisted var episodeId: String?
+    @Persisted var duration: Double = 0
+    @Persisted var progress: Double = 0
+    @Persisted var currentTime: Double = 0
+    @Persisted var isFinished: Bool = false
+    @Persisted var lastUpdate: Int = 0
+    @Persisted var startedAt: Int = 0
+    @Persisted var finishedAt: Int?
+    
+    private enum CodingKeys : String, CodingKey {
+        case id, libraryItemId, episodeId, duration, progress, currentTime, isFinished, lastUpdate, startedAt, finishedAt
+    }
+    
+    override init() {
+        super.init()
+    }
+    
+    required init(from decoder: Decoder) throws {
+        let values = try decoder.container(keyedBy: CodingKeys.self)
+        id = try values.decode(String.self, forKey: .id)
+        libraryItemId = try values.decode(String.self, forKey: .libraryItemId)
+        episodeId = try? values.decode(String.self, forKey: .episodeId)
+        duration = try values.doubleOrStringDecoder(key: .duration)
+        progress = try values.doubleOrStringDecoder(key: .progress)
+        currentTime = try values.doubleOrStringDecoder(key: .currentTime)
+        isFinished = try values.decode(Bool.self, forKey: .isFinished)
+        lastUpdate = try values.intOrStringDecoder(key: .lastUpdate)
+        startedAt = try values.intOrStringDecoder(key: .startedAt)
+        finishedAt = try? values.intOrStringDecoder(key: .finishedAt)
+    }
+    
+    func encode(to encoder: Encoder) throws {
+        var container = encoder.container(keyedBy: CodingKeys.self)
+        try container.encode(id, forKey: .id)
+        try container.encode(libraryItemId, forKey: .libraryItemId)
+        try container.encode(episodeId, forKey: .episodeId)
+        try container.encode(duration, forKey: .duration)
+        try container.encode(progress, forKey: .progress)
+        try container.encode(currentTime, forKey: .currentTime)
+        try container.encode(isFinished, forKey: .isFinished)
+        try container.encode(lastUpdate, forKey: .lastUpdate)
+        try container.encode(startedAt, forKey: .startedAt)
+        try container.encode(finishedAt, forKey: .finishedAt)
+    }
+}
